% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn-ssf.R
\name{cnn_ssf}
\alias{cnn_ssf}
\title{CNN-SSF Method for Data Frames}
\usage{
cnn_ssf(X, Y, density = NULL, batchsize = 0.1, n_control, ...)
}
\arguments{
\item{X}{Array of images describing the habitat}

\item{Y}{Vectors of steps and pseudo steps (1 and 0)}

\item{density}{density of steps given the movement kenrnel that was used to sample the random steps}

\item{batchsize}{Batch size (in percent of the data) used during neural network training.}

\item{n_control}{Number of pseudo/random steps}

\item{...}{Arguments passed on to \code{\link[cito:cnn]{cito::cnn()}}}
}
\value{
A list of class \code{"cnn_ssf"} containing the fitted model and metadata.
}
\description{
Fits a convolutional neural network step-selection model from a \code{data.frame}
containing movement steps, covariates, and case/control indicators.
}
\examples{
library(cito)
library(citoMove)
n = 100*11
X = array(runif(n*5*5*1, -1, 1), dim = c(n, 5, 5))
softmax = function(x) exp(x)/sum(exp(x))
YY = apply(matrix(sapply(1:n, function(i) sum((X[i,,] * diag(1, 5, 5)))*2 ), nrow = 11), 2, function(x) rmultinom(1, 1, prob = softmax(x)))
Y = as.vector(YY)

X = array(X, dim = c(n, 1, 5, 5))
architecture = create_architecture(conv(n_kernels = 1L, kernel_size = c(5,5)), linear(20, activation = "relu"))
m = cnn_ssf(X = X, Y = Y, density = rep(1, n), n_control = 10L, architecture= architecture, lr = 0.1, epochs = 500L)
}
\seealso{
\code{\link[cito:cnn]{cito::cnn()}}
}
