% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn-ssf.R
\name{dnn_ssf.random_steps}
\alias{dnn_ssf.random_steps}
\title{DNN-SSF Method for Data Frames}
\usage{
\method{dnn_ssf}{random_steps}(data, formula, batchsize = "auto", ...)
}
\arguments{
\item{data}{object of class \code{random_steps} from the \code{amt} package.}

\item{formula}{A model formula of the form \code{case_ ~ cov1 + cov2 + ...}
specifying the model to be fit.}

\item{batchsize}{Batch size used during neural network training. Either
\code{"auto"} (default) or a numeric value.}

\item{...}{
  Arguments passed on to \code{\link[=dnn_ssf]{dnn_ssf}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A list of class \code{"dnn_ssf"} containing the fitted model and metadata.
}
\description{
Fits a deep neural network step-selection model from a \code{data.frame}
containing movement steps, covariates, and case/control indicators.
}
\examples{
fit <- dnn_ssf(
   data = mydata,
  formula = case_ ~ sl_ + hab,
  batchsize = 64
 )
summary(fit)
predict(fit, newdata = mydata)


library(cito)
?cito
?citoMove
}
\seealso{
\code{\link[=dnn_ssf]{dnn_ssf()}}, \code{\link[cito:dnn]{cito::dnn()}}
}
