% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn-ssf.R
\name{dnn_ssf}
\alias{dnn_ssf}
\title{Fit a Deep Neural Network Step-Selection Function (dnn_ssf)}
\usage{
dnn_ssf(data, ...)
}
\arguments{
\item{data}{object of class \code{random_steps} from the \code{amt} package.}
}
\description{
Fits a custom deep neural network using \code{\link{dnn}}.
}
