% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citoMove.R
\docType{package}
\name{citoMove}
\alias{citoMove}
\alias{citoMove-package}
\title{citoMove - an R package for animal movement analysis using deep learning}
\description{
The 'citoMove' package builds upon the the step selection function framework as well as deep learning algorithms and is able to predict animals movement and behavioral states based on their physical and social environmental variables.
}
\details{
See index / vignette for details
}
\examples{
vignette("citoMove", package="citoMove")
}
\seealso{
\code{\link{fitDeepSSF}}
}
\author{
\strong{Maintainer}: Maximilian Pichler \email{maximilian.pichler@biologie.uni-regensburg.de} (\href{https://orcid.org/0000-0003-2252-8327}{ORCID})

Authors:
\itemize{
  \item Thibault Fronville
  \item Florian Hartig (\href{https://orcid.org/0000-0002-6255-9059}{ORCID})
  \item Johannes Signer
}

}
\keyword{internal}
