% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnn-ssf.R
\name{dnn_ssf}
\alias{dnn_ssf}
\title{DNN-SSF Method for Data Frames}
\usage{
dnn_ssf(formula, data, batchsize = 0.1, n_control = NULL, ...)
}
\arguments{
\item{formula}{A model formula of the form \code{case_ ~ cov1 + cov2 + ...}
specifying the model to be fit.}

\item{data}{Data.frame containing the true and random steps, and the covariates.}

\item{batchsize}{Batch size (in percent of the data) used during neural network training.}

\item{n_control}{Number of pseudo absences.}

\item{...}{Arguments passed on to \code{\link[cito:dnn]{cito::dnn()}}.}
}
\value{
A list of class \code{"dnn_ssf"} containing the fitted model and metadata.
}
\description{
Fits a deep neural network step-selection model from a \code{data.frame}
containing movement steps, covariates, and case/control indicators.
}
\examples{
library(amt)
library(citoMove)
library(terra)

data(wildboar)
str(wildboar)

# create a new covariate: the distance to the next forest.

forest <- amt::get_sh_forest()
forest <- terra::subst(forest, 0, NA)
forest_dist <- distance(forest)
names(forest_dist) <- "forest_dist"
plot(forest_dist)

# Load the tracking data of a red deer, create random steps, extract covariate

data(deer)
dat_ssf <- deer |>
  steps_by_burst() |>
  random_steps() |>
  extract_covariates(forest_dist) |>
  time_of_day() |>
  mutate(case_ = as.integer(case_))


# Fit a deep neural network (dnn):

model = dnn_ssf(case_ ~ forest_dist, data = dat_ssf, epoch = 10L, plot = FALSE, verbose = FALSE)
summary(model)
}
\seealso{
\code{\link[cito:dnn]{cito::dnn()}}
}
