% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plotSpatRasterCollection}
\alias{plotSpatRasterCollection}
\title{Plot SpatRasterCollection}
\usage{
plotSpatRasterCollection(x, colours, onlyoverlap = TRUE)
}
\arguments{
\item{x}{object of class 'SpatRasterCollection'}

\item{colours}{vector of colours of the layers. Must be same length as number of layers}

\item{onlyoverlap}{if TRUE only overlap of all layers will be plotted. If FALSE whole layers will be plotted}
}
\description{
Plots an object of class 'SpatRasterCollection'
}
\examples{

#' To simulate a track we need an object of class SpatRasterCollection
#' The object has to obtain one or multiple objects of class SpatRaster


#' Example for simulating two raster layers with the terra package, combining them in an object of SpatRasterCollection and plotting them with the plotSpatRasterCollection function:

#' The first layer is a layer with a continuous variable:
layer1 <- terra::rast(ncols = 300, nrows = 450,
                      ext = terra::ext(-100, 100, -150, 150))     # defining extend and resolution of the layer with resolution being number of cells per map unit
terra::values(layer1) <- stats::rnorm(terra::ncell(layer1))
layer1 <- terra::focal(layer1, w = terra::focalMat(layer1, 5, "Gauss"),
                       na.policy = "omit", expand = TRUE)         # sampling values for the cells and smoothening them with terra::focalMat function

terra::plot(layer1,
            col = adjustcolor(colorRampPalette(c("white", "darkblue"))(100), alpha = 1/3),
            legend = FALSE)     # plot (just for visualization)


#' The second raster is a layer with a binary variable:
layer2 <- terra::rast(ncols = 300, nrows = 200,
                      ext = terra::ext(-150, 150, -100, 100))
terra::values(layer2) <- stats::rnorm(terra::ncell(layer2))
layer2 <- terra::focal(layer2, w = terra::focalMat(layer2, 15, "Gauss"),
                       na.policy = "omit", expand = TRUE)
threshold <- stats::quantile(terra::values(layer2), 0.5, na.rm=TRUE)
layer2 <- layer2 >= threshold     # assigning and implementing a threshold for binary values

terra::plot(layer2,
            col = c(adjustcolor(NA, alpha = 0), adjustcolor("darkgreen", alpha = 1/3)),
            legend = FALSE)   # plot (just for visualization)


#' The two layers of class SpatRaster are combined in an object of class SpatRasterCollection
layers <- terra::sprc(layer1, layer2)


#' The SpatrasterCollection is plotted with this function. Showing only the area where all layers overlap when onlyoverlap = TRUE or the whole rasters when onlyoverlap = FALSE
plotSpatRasterCollection(layers,
                         colours = c("darkblue", "darkgreen"),
                         onlyoverlap = FALSE)


#' The track is simulated with this function.
#'   - numericLayers must be an object of class SpatRasterCollection containing one or multiple SpatRaster.
#'   - betas must be a vector of preference values for each of the layers. Must be same length as number of layers
track1 <- simulateTrack(numericLayers = layers,
                        betas = c(3, -5))


#' To visualize the plot the points and/or lines function can be used, adding the track to a beforehand plotted layer or SpatRastercollection. Other tracks of the amt class track can be added to the plot in the same way.
points(track1$x_[1], track1$y_[1], pch = 16, col = "black")
lines(track1$x_, track1$y_, col = "black")

}
\author{
Mika Schubert
}
