% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildboar.R
\docType{data}
\name{wildboar}
\alias{wildboar}
\title{Prepocessed data of tracked wild boards}
\format{
A tibble with 68,565 rows and 11 variables:
\describe{
\item{id}{Factor identifying each wild boar (8 individuals).}
\item{sl_}{Numeric Step length.}
\item{ta_}{Numeric Turning angle between consecutive steps (radians).}
\item{t1_}{POSIXct Start timestamp of the step.}
\item{t2_}{POSIXct End timestamp of the step.}
\item{dt_}{Difftime (hours). Time interval between \code{t1_} and \code{t2_} (all equal to 12 hours).}
\item{case_}{Logical. \code{TRUE} indicates an observed step; \code{FALSE} indicates a control (available) step.}
\item{step_id_}{Character. Identifier linking observed steps to their matched control steps.}
\item{dtw}{Numeric. Distance to water.}
\item{imp}{Numeric. Environmental or model-based importance score.}
\item{hab}{Numeric. Habitat type or indicator variable.}
}
}
\source{
Stillfried, M., Gras, P., Börner, K., Göritz, F., Painer, J., Röllig, K.,
Wenzler, M., Hofer, H., Ortmann, S., & Kramer-Schadt, S. (2017).
\emph{Secrets of Success in a Landscape of Fear: Urban Wild Boar Adjust Risk
Perception and Tolerate Disturbance}. \emph{Frontiers in Ecology and Evolution},
5:157. https://doi.org/10.3389/fevo.2017.00157. :contentReference[oaicite:3]{index=3}
}
\usage{
wildboar
}
\description{
A dataset containing observed and random steps annotated environmental covariates
for GPS-tracked wild boars. The GPS data recorded in a study
of urban and rural wild boar movement patterns in the Berlin area.
}
\examples{
library(amt)
library(citoMove)
library(terra)

data(wildboar)
str(wildboar)

# create a new covariate: the distance to the next forest.

forest <- amt::get_sh_forest()
forest <- terra::subst(forest, 0, NA)
forest_dist <- distance(forest)
names(forest_dist) <- "forest_dist"
plot(forest_dist)

# Load the tracking data of a red deer, create random steps, extract covariate

data(deer)
dat_ssf <- deer |>
  steps_by_burst() |>
  random_steps() |>
  extract_covariates(forest_dist) |>
  time_of_day() |>
  mutate(case_ = as.integer(case_))


# Fit a deep neural network (dnn):

model = dnn_ssf(case_ ~ forest_dist, data = dat_ssf, epoch = 10L, plot = FALSE, verbose = FALSE)
summary(model)


}
\keyword{datasets}
